Tools via mod, update protoc-gen-go, support new protoc
Wanted to update protoc-gen-go to match the protobuf update. Couldn't get
retool to work properly, despite trying several older versions of Go. The
project is also archived, so tool the time to switch to simple, Go modules
based installs of the tools. This is done without actually changing this
library to need modules, support is only flagged in explicitly to install
those. This is used to install the version of protoc-gen-go that matches the
vendored copy, and the latest version of errcheck is installed.

The protoc check script is updated - the versioning scheme has changed, so we
now just check a minimum of protoc 3.

Generated pb files are re-generated.